+++ CHANGELOG.md
@@ -4,6 +4,11 @@
 All notable changes to this project (as seen by library users) will be documented in this file.
 The CHANGELOG is available on [Github](https://github.com/luc-tielen/souffle-haskell.git/CHANGELOG.md).
 
+## [???] - Unreleased
+### Added
+
+- getFacts can now return facts in a fixed size array (from the array package).
+
 ## [1.0.0] - 2020-07-09
 ### Changed